requirements.txt
gradio>=4.0.0
dashscope>=1.14.0
openai>=1.0.0
pydantic>=2.0.0
json-repair>=0.25.0
python-dotenv>=1.0.0
Pillow>=10.0.0
